pyproject.toml
[project]
name = "games"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pygame-ce>=2.5.7",
]

[project.scripts]
games = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"



[tool.uv]
package = true